Extract IConnection and NetworkAddress out from network.h
