+++ fdbclient/Tracing.actor.cpp
@@ -23,6 +23,7 @@
 #include "flow/IRandom.h"
 #include "flow/UnitTest.h"
 #include "flow/Knobs.h"
+#include "flow/IConnection.h"
 #include "fdbclient/IKnobCollection.h"
 #include "flow/network.h"
 #include <functional>
