+++ fdbclient/include/fdbclient/S3BlobStore.h
@@ -29,6 +29,8 @@
 #include "fdbrpc/HTTP.h"
 #include "fdbclient/JSONDoc.h"
 
+class IConnection;
+
 // Representation of all the things you need to connect to a blob store instance with some credentials.
 // Reference counted because a very large number of them could be needed.
 class S3BlobStoreEndpoint : public ReferenceCounted<S3BlobStoreEndpoint> {
