+++ fdbserver/MetricLogger.actor.cpp
@@ -41,6 +41,7 @@
 #include "flow/flow.h"
 #include "flow/network.h"
 #include "flow/IUDPSocket.h"
+#include "flow/IConnection.h"
 #include "flow/actorcompiler.h" // This must be the last #include.
 
 struct MetricsRule {
