+++ fdbserver/workloads/UDPWorkload.actor.cpp
@@ -35,7 +35,7 @@
 #include <memory>
 #include <functional>
 #include "flow/IUDPSocket.h"
-
+#include "flow/IConnection.h"
 #include "flow/actorcompiler.h" // has to be last include
 
 namespace {
