+++ flow/Hostname.actor.cpp
@@ -22,6 +22,7 @@
 
 #include <regex>
 
+#include "flow/IConnection.h"
 #include "flow/UnitTest.h"
 
 #include "flow/actorcompiler.h" // has to be last include
